Exrtaction des classements dans une composans externes
